Added Boolean array flag_empty_in, that is used for masking the output array. This is used to check which cells do not contain VDF data already, avoiding unecessary distance computations.
Tested that the vectorized and loop versions give the same output.
Tested that the function works quickly on a restart file, in which every cell was populated with a vdf.
Loop version could be very slow (many minutes) for certain use cases, such as finding the nearest vdf-containing cellID for every vg cell in the simulation.